requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Werkzeug==3.0.1
python-dotenv==1.0.1
supabase==2.3.1
pandas==2.2.1
openpyxl==3.1.2
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2024.1
numpy
matplotlib
seaborn
scipy
xlsxwriter 